requirements.txt
Flask==2.3.2
gspread==5.9.0
google-auth==2.23.3
aiohttp==3.9.0b0
